Revert proof model file separation
