+++ open_prices/proofs/ml.py
@@ -7,7 +7,7 @@ from openfoodfacts.ml.object_detection import ObjectDetectionRawResult, ObjectDe
 from PIL import Image
 
 from . import constants
-from .models.proof import Proof, ProofPrediction
+from .models import Proof, ProofPrediction
 
 logger = logging.getLogger(__name__)
